src/main/kotlin/io/github/cianciustyles/instructions/Load.kt
package io.github.cianciustyles.instructions

import io.github.cianciustyles.LC3VM
import io.github.cianciustyles.Utils.extendSign

@ExperimentalUnsignedTypes
class Load(encoding: UShort) : Instruction() {
    val destinationRegister: UShort = (encoding.toInt() shr 9 and 0x7).toUShort()
    val pcOffset: Short = extendSign(encoding.toInt() and 0x1FF, 9)

    override fun execute(vm: LC3VM) =
        loadAndSetConditionCodes(
            vm,
            (vm.registers.getPC() + pcOffset).toUShort(),
            destinationRegister
        )
}
